refactor: bench folder uses binaries instead of CLI
